% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alonso15.R
\name{alonso15}
\alias{alonso15}
\title{Lakshadweep Archipelago coral fish community reassembly}
\format{A list with 3 dataframes, each corresponding to the survey of a
  different atoll. Dataframes have in columns: \describe{ \item{Species}{Name
  of the species found} \item{Trophic.Level}{A number indicating the trophic
  level of the surveyed species} \item{Presence-absence data}{Several columns
  with letters (indicating the atoll surveyed) and the year in which the
  surveys were done} \item{Guild}{Guild of the surveyed species} }}
\source{
Alonso, D., Pinyol-Gallemi, A., Alcoverro T. and Arthur, R.. (2015)
  Fish community reassembly after a coral mass mortality: higher trophic
  groups are subject to increased rates of extinction. \emph{Ecology
  Letters}, \bold{18}, 451--461.
}
\description{
A list with three datasets containing presence-absence data for the
reassembly proccess of coral fish communities in three atolls (Agatti, Kadmat
and Kavaratti) of the Lakshadweep Archipelago (India).
}
\details{
Surveys were conducted from 2000 to 2011 in order to follow community
reassembly after a coral mass mortality event in the relatively unfished
Lakshadweep Archipelago.  Results indicated that higher trophic groups suffer
an increased extinction rate even without fishing targeting them.
}
\note{
Kavaratti atoll was not surveyed in 2000 and 2010.
}
